+++ source/extensions/omni.isaac.core_nodes/premake5.lua
@@ -35,6 +35,7 @@ project_ext_plugin(ext, "omni.isaac.core_nodes.plugin")
         "%{root}/_build/target-deps/client_library/include",
         "%{kit_sdk_bin_dir}/extscore/omni.syntheticdata/include",
         "%{root}/_build/kit_%{config}/_exts/omni.syntheticdata/include",
+        "%{root}/_build/target-deps/usd_ext_isaac/%{cfg.buildcfg}/include",
      }
      libdirs {
         "%{root}/_build/target-deps/nv_usd/%{cfg.buildcfg}/lib",
@@ -46,7 +47,7 @@ project_ext_plugin(ext, "omni.isaac.core_nodes.plugin")
     }
 
      links {
-        "gf", "sdf", "tf", "usd", "usdGeom", "usdUtils", "omni.usd",
+        "gf", "sdf", "tf", "usd", "usdGeom", "usdUtils", "omni.usd", "rangeSensorSchema",
     }
 
     filter { "configurations:debug" }