+++ source/extensions/omni.isaac.ros_bridge/nodes/OgnROS1PublishLaserScan.cpp
@@ -11,14 +11,9 @@
 #include <UsdPCH.h>
 // clang-format on
 
-#include "omni/isaac/utils/UsdUtilities.h"
 #include "sensor_msgs/LaserScan.h"
 
-#include <carb/flatcache/FlatCache.h>
-
-#include <omni/isaac/range_sensor/RangeSensorInterface.h>
 #include <omni/isaac/ros/RosNode.h>
-#include <rangeSensorSchema/lidar.h>
 
 #include <OgnROS1PublishLaserScanDatabase.h>
 
@@ -26,24 +21,13 @@
 class OgnROS1PublishLaserScan : public RosNode
 {
 public:
-    static void initialize(const GraphContextObj& contextObj, const NodeObj& nodeObj)
-    {
-        auto& state = OgnROS1PublishLaserScanDatabase::sInternalState<OgnROS1PublishLaserScan>(nodeObj);
-
-        state.mLidarSensorInterface = carb::getCachedInterface<omni::isaac::range_sensor::LidarSensorInterface>();
-
-        if (!state.mLidarSensorInterface)
-        {
-            CARB_LOG_ERROR("Failed to acquire omni::isaac::range_sensor interface");
-            return;
-        }
-    }
+    // static void initialize(const GraphContextObj& contextObj, const NodeObj& nodeObj)
+    // {
+    //     auto& state = OgnROS1PublishLaserScanDatabase::sInternalState<OgnROS1PublishLaserScan>(nodeObj);
+    // }
 
     static bool compute(OgnROS1PublishLaserScanDatabase& db)
     {
-        const GraphContextObj& context = db.abi_context();
-
-
         auto& state = db.internalState<OgnROS1PublishLaserScan>();
 
         // spin once calls reset automatically if it was not successful
@@ -56,33 +40,6 @@ public:
         // Publisher was not valid, create a new one
         if (!state.mPublisher)
         {
-            const char* primPath = db.inputs.lidarPrim.path();
-
-            // Find our stage
-            long stageId = context.iContext->getStageId(context);
-            auto stage = pxr::UsdUtilsStageCache::Get().Find(pxr::UsdStageCache::Id::FromLongInt(stageId));
-            if (!stage)
-            {
-                db.logError("Could not find USD stage %ld", stageId);
-                return false;
-            }
-
-            // Verify we have a valid lidar prim
-            pxr::UsdPrim targetPrim = stage->GetPrimAtPath(pxr::SdfPath(primPath));
-            if (!targetPrim.IsA<pxr::RangeSensorSchemaLidar>())
-            {
-                db.logError("Prim is not a Lidar Prim");
-                return false;
-            }
-
-            state.mLidarPrim = pxr::RangeSensorSchemaLidar(targetPrim);
-
-            if (!state.mLidarSensorInterface->isLidarSensor(primPath))
-            {
-                db.logError("Prim is not registered with Lidar extension");
-                return false;
-            }
-
             // Setup ROS publisher
             const std::string& topicName = db.inputs.topicName();
 
@@ -93,8 +50,6 @@ public:
 
             state.mPublisher = std::make_unique<ros::Publisher>(
                 state.mNodeHandle->advertise<sensor_msgs::LaserScan>(topicName, db.inputs.queueSize()));
-
-            state.mLidarPrimPath = primPath;
 
             state.mFrameId = db.inputs.frameId();
             addFramePrefix(db.inputs.nodeNamespace(), state.mFrameId);
@@ -110,182 +65,67 @@ public:
     void publishLidar(OgnROS1PublishLaserScanDatabase& db)
     {
         CARB_PROFILE_ZONE(0, "Lidar 2D Pub");
-        if (!mLidarSensorInterface->isLidarSensor(mLidarPrimPath))
-        {
-            db.logError("Invalid Lidar Reference, Prim is not registered with Lidar extension");
-            return;
-        }
 
-        bool highLod = false;
-        omni::isaac::utils::safeGetAttribute(mLidarPrim.GetHighLodAttr(), highLod);
-
-        if (highLod)
-        {
-            db.logError(
-                "High LOD not supported for LaserScan, only 2D Lidar Supported for LaserScan. Please disable Lidar High LOD setting or uncheck LaserScanEnabled");
-            return;
-        }
+        // Setup ROS Lidar Message
         sensor_msgs::LaserScan laser_msg;
         laser_msg.header.seq = 0;
         laser_msg.header.frame_id = mFrameId;
 
-        int numColsTicked = mLidarSensorInterface->getNumColsTicked(mLidarPrimPath);
-        int numRows = mLidarSensorInterface->getNumRows(mLidarPrimPath); // should be 1
-        if (numRows > 1)
+        if (db.inputs.numRows() != 1)
         {
-            db.logError("High LOD not supported for LaserScan, only 2D Lidar Supported for LaserScan");
-        }
-        size_t numBeams = numColsTicked * numRows;
-
-        float* theta = mLidarSensorInterface->getAzimuthData(mLidarPrimPath);
-        // float* phi = mLidarSensorInterface->getZenithData(mLidarPrimPath); // should have one entry
-        float* ranges = mLidarSensorInterface->getLinearDepthData(mLidarPrimPath);
-        uint8_t* intensities = mLidarSensorInterface->getIntensityData(mLidarPrimPath);
-
-        if (!theta || !ranges || !intensities)
-        {
+            db.logError(
+                "Number of rows must be equal to 1. High LOD not supported for LaserScan, only 2D Lidar Supported for LaserScan. Please disable Lidar High LOD setting");
             return;
         }
 
-        float maxRange = 100;
-        float minRange = 0.4;
-        float rotationRate = 0.0;
-        float horizontalResolution = 0.4;
-        float horizontalFov = 360.0;
-
-        omni::isaac::utils::safeGetAttribute(mLidarPrim.GetMaxRangeAttr(), maxRange);
-        omni::isaac::utils::safeGetAttribute(mLidarPrim.GetMinRangeAttr(), minRange);
-        omni::isaac::utils::safeGetAttribute(mLidarPrim.GetRotationRateAttr(), rotationRate);
-        omni::isaac::utils::safeGetAttribute(mLidarPrim.GetHorizontalResolutionAttr(), horizontalResolution);
-        omni::isaac::utils::safeGetAttribute(mLidarPrim.GetHorizontalFovAttr(), horizontalFov);
-
-        size_t numBeamsTotal = horizontalFov / horizontalResolution;
-
-        if (horizontalResolution == 0.0)
+        if (db.inputs.timeStamp() >= 0.0)
         {
-            db.logError("Lidar Prim %s: Horizontal Resolution must be greater than 0.0", mLidarPrimPath);
-            return;
+            laser_msg.header.stamp.fromSec(db.inputs.timeStamp());
         }
-        if (horizontalFov == 0.0)
+        else
         {
-            db.logError("Lidar Prim %s: Horizontal FOV must be greater than 0.0", mLidarPrimPath);
-            return;
+            db.logWarning("Timestamp is invalid. Timestamp will be neglected for all published ROS LaserScan messages");
         }
-        if (horizontalFov > 360.0)
+
+        laser_msg.angle_min = db.inputs.azimuthRange()[0];
+        laser_msg.angle_max = db.inputs.azimuthRange()[1];
+
+        float rotationRate = db.inputs.rotationRate();
+        laser_msg.scan_time = rotationRate ? 1.0 / rotationRate : 0;
+        laser_msg.range_min = db.inputs.depthRange()[0];
+        laser_msg.range_max = db.inputs.depthRange()[1];
+
+        size_t buffSize = db.inputs.numCols() * db.inputs.numRows();
+
+        if (!db.inputs.linearDepthData.isValid() || !db.inputs.intensitiesData.isValid())
         {
-            db.logError("Lidar Prim %s: Horizontal FOV must be less than or equal to 360.0", mLidarPrimPath);
+            db.logError("Buffers are invalid");
             return;
         }
 
-        uint64_t curr_sequence_num = mLidarSensorInterface->getSequenceNumber(mLidarPrimPath);
-
-        if (curr_sequence_num < mPrevSequenceNumber)
+        if (db.inputs.linearDepthData.size() != db.inputs.intensitiesData.size())
         {
-            mResetLaserScan = true;
-            mPrevSequenceNumber = curr_sequence_num;
+            db.logError("Linear Depth data and Intensities data sizes do not match");
+            return;
         }
 
-        carb::Float2 azimuthRange = mLidarSensorInterface->getAzimuthRange(mLidarPrimPath);
-
-        if (mResetLaserScan)
+        if (buffSize != db.inputs.linearDepthData.size())
         {
-            mIntensitiesData.clear();
-            mRangesData.clear();
-            mNumBeamsRemaining = numBeamsTotal;
-            mPrevRotationRate = rotationRate;
-            mPrevHorizontalResolution = horizontalResolution;
-            mPrevHorizontalFov = horizontalFov;
-
-            bool foundStart = false;
-            for (mBeamIdx = 0; mBeamIdx < numBeams; mBeamIdx++)
-            {
-                if (theta[mBeamIdx] == azimuthRange.x)
-                {
-                    foundStart = true;
-                    break;
-                }
-            }
-            if (!foundStart)
-            {
-                return;
-            }
-            mResetLaserScan = false;
+            db.logError("Lidar data with %d rows and %d columns does not match input buffer array size of %d",
+                        db.inputs.numRows(), db.inputs.numCols(), db.inputs.linearDepthData.size());
+            return;
         }
 
-        if (mNumBeamsRemaining > numBeams)
-        {
-            for (size_t i = mBeamIdx; i < numBeams; i++)
-            {
-                mIntensitiesData.push_back(static_cast<float>(intensities[i]));
-                mRangesData.push_back(ranges[i]);
-                mNumBeamsRemaining--;
-            }
-            mBeamIdx = 0;
-        }
+        laser_msg.ranges.resize(buffSize);
+        laser_msg.ranges.assign(db.inputs.linearDepthData().begin(), db.inputs.linearDepthData().end());
 
-        else if (mNumBeamsRemaining <= numBeams)
-        {
+        laser_msg.intensities.resize(buffSize);
+        laser_msg.intensities.assign(db.inputs.intensitiesData().begin(), db.inputs.intensitiesData().end());
 
-            // Save data up to maximum FOV
-            size_t idx;
-            for (idx = 0; idx < mNumBeamsRemaining; idx++)
-            {
-                mIntensitiesData.push_back(static_cast<float>(intensities[idx]));
-                mRangesData.push_back(ranges[idx]);
-            }
+        laser_msg.angle_increment = db.inputs.horizontalResolution() * M_PI / 180.0;
+        laser_msg.time_increment = (db.inputs.horizontalFov() / 360.0 * laser_msg.scan_time) / laser_msg.ranges.size();
 
-            // Setup ROS Lidar Message
-
-            if (db.inputs.timeStamp() >= 0.0)
-            {
-                laser_msg.header.stamp.fromSec(db.inputs.timeStamp());
-            }
-            else
-            {
-                db.logWarning(
-                    "Timestamp is invalid. Timestamp will be neglected for all published ROS LaserScan messages");
-            }
-
-            laser_msg.angle_min = azimuthRange.x;
-            laser_msg.angle_max = azimuthRange.y;
-
-            laser_msg.scan_time = rotationRate ? 1.0 / rotationRate : 0;
-            laser_msg.range_min = minRange;
-            laser_msg.range_max = maxRange;
-
-            laser_msg.ranges = mRangesData;
-            laser_msg.intensities = mIntensitiesData;
-
-            laser_msg.angle_increment = horizontalResolution * M_PI / 180.0;
-            laser_msg.time_increment = (horizontalFov / 360.0 * laser_msg.scan_time) / laser_msg.ranges.size();
-
-            mPublisher->publish(laser_msg);
-
-            mPrevSequenceNumber = curr_sequence_num;
-
-            // Reset fields for new ROS Lidar message
-            mRangesData.clear();
-            mIntensitiesData.clear();
-
-            if (idx < numBeams)
-            {
-                if (theta[idx] != azimuthRange.x)
-                {
-                    mResetLaserScan = true;
-                    return;
-                }
-            }
-
-            // Save remaining data
-            size_t numBeamsOffset = numBeams - mNumBeamsRemaining;
-            for (size_t j = 0; j < numBeamsOffset; j++)
-            {
-                mIntensitiesData.push_back(static_cast<float>(intensities[idx]));
-                mRangesData.push_back(ranges[idx]);
-                idx++;
-            }
-            mNumBeamsRemaining = numBeamsTotal - numBeamsOffset;
-        }
+        mPublisher->publish(laser_msg);
     }
 
     virtual void release(const NodeObj& nodeObj)
@@ -296,7 +136,6 @@ public:
 
     virtual void reset()
     {
-        mResetLaserScan = true;
         mPublisher.reset(); // This should be reset before we reset the handle.
         RosNode::reset();
     }
@@ -305,25 +144,7 @@ public:
 private:
     std::unique_ptr<ros::Publisher> mPublisher;
 
-    omni::isaac::range_sensor::LidarSensorInterface* mLidarSensorInterface = nullptr;
-    pxr::RangeSensorSchemaLidar mLidarPrim;
-
-    const char* mLidarPrimPath = nullptr;
-
     std::string mFrameId = "sim_lidar";
-    std::vector<float> mIntensitiesData;
-    std::vector<float> mRangesData;
-
-    uint64_t mPrevSequenceNumber = 0;
-
-    bool mResetLaserScan = true;
-    size_t mNumBeamsRemaining;
-
-    float mPrevRotationRate;
-    float mPrevHorizontalResolution;
-    float mPrevHorizontalFov;
-
-    size_t mBeamIdx = 0;
 };
 
 REGISTER_OGN_NODE()